requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coverage==5.5
dataclasses==0.8
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.1.0
pkg-resources==0.0.0
psycopg2-binary==2.8.5
PyYAML==5.4.1
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.17
typing-extensions==3.7.4.3
urllib3==1.25.9
vcrpy==4.1.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
